Add some modifications to PR
