+++ firebase-auth/src/main/java/com/androidhuman/rxfirebase2/auth/CreateUserWithEmailAndPasswordObserver.java
@@ -36,7 +36,7 @@ class CreateUserWithEmailAndPasswordObserver extends Single<FirebaseUser> {
                 .addOnCompleteListener(listener);
     }
 
-    static final class Listener extends OnCompleteDisposable<AuthResult> {
+    private static final class Listener extends OnCompleteDisposable<AuthResult> {
 
         private final SingleObserver<? super FirebaseUser> observer;
