+++ firebase-auth/src/main/java/com/androidhuman/rxfirebase2/auth/RxFirebaseAuth.java
@@ -95,11 +95,11 @@ public final class RxFirebaseAuth {
             @NonNull FirebaseAuth instance) {
         return new SignInAnonymouslyObserver(instance)
                 .map(new Function<AuthResult, FirebaseUser>() {
-            @Override
-            public FirebaseUser apply(AuthResult authResult) throws Exception {
-                return authResult.getUser();
-            }
-        });
+                    @Override
+                    public FirebaseUser apply(AuthResult authResult) throws Exception {
+                        return authResult.getUser();
+                    }
+                });
     }
 
     @CheckResult
