+++ firebase-auth/src/main/java/com/androidhuman/rxfirebase2/auth/SignInWithCustomTokenObserver.java
@@ -1,7 +1,12 @@
 package com.androidhuman.rxfirebase2.auth;
 
+import com.google.android.gms.tasks.Task;
 import com.google.firebase.auth.AuthResult;
 import com.google.firebase.auth.FirebaseAuth;
+
+import com.androidhuman.rxfirebase2.core.OnCompleteDisposable;
+
+import android.support.annotation.NonNull;
 
 import io.reactivex.Single;
 import io.reactivex.SingleObserver;
@@ -19,10 +24,30 @@ final class SignInWithCustomTokenObserver extends Single<AuthResult> {
 
     @Override
     protected void subscribeActual(SingleObserver<? super AuthResult> observer) {
-        AuthResultListener listener = new AuthResultListener(observer);
+        Listener listener = new Listener(observer);
         observer.onSubscribe(listener);
 
         instance.signInWithCustomToken(token)
                 .addOnCompleteListener(listener);
+    }
+
+    private final class Listener extends OnCompleteDisposable<AuthResult> {
+
+        private final SingleObserver<? super AuthResult> observer;
+
+        Listener(@NonNull SingleObserver<? super AuthResult> observer) {
+            this.observer = observer;
+        }
+
+        @Override
+        public void onComplete(@NonNull Task<AuthResult> task) {
+            if (!isDisposed()) {
+                if (!task.isSuccessful()) {
+                    observer.onError(task.getException());
+                } else {
+                    observer.onSuccess(task.getResult());
+                }
+            }
+        }
     }
 }
