+++ firebase-auth/src/main/java/com/androidhuman/rxfirebase2/auth/UpdateCurrentUserObserver.java
@@ -32,7 +32,7 @@ final class UpdateCurrentUserObserver extends Completable {
                 .addOnCompleteListener(listener);
     }
 
-    static final class Listener extends OnCompleteDisposable<Void> {
+    private static final class Listener extends OnCompleteDisposable<Void> {
 
         private final CompletableObserver observer;
