+++ firebase-auth/src/main/java/com/androidhuman/rxfirebase2/auth/UserUpdatePasswordObserver.java
@@ -30,7 +30,7 @@ final class UserUpdatePasswordObserver extends Completable {
                 .addOnCompleteListener(listener);
     }
 
-    static final class Listener extends OnCompleteDisposable<Void> {
+    private static final class Listener extends OnCompleteDisposable<Void> {
 
         private final CompletableObserver observer;
