+++ firebase-auth/src/main/java/com/androidhuman/rxfirebase2/auth/UserUpdatePhoneNumberObserver.java
@@ -31,7 +31,7 @@ final class UserUpdatePhoneNumberObserver extends Completable {
                 .addOnCompleteListener(listener);
     }
 
-    static final class Listener extends OnCompleteDisposable<Void> {
+    private static final class Listener extends OnCompleteDisposable<Void> {
 
         private final CompletableObserver observer;
