+++ firebase-auth/src/test/kotlin/com/androidhuman/rxfirebase2/auth/RxFirebaseAuthTest.kt
@@ -2,8 +2,17 @@ package com.androidhuman.rxfirebase2.auth
 
 import com.google.android.gms.tasks.OnCompleteListener
 import com.google.android.gms.tasks.Task
-import com.google.firebase.auth.*
-import com.nhaarman.mockito_kotlin.*
+import com.google.firebase.auth.AuthCredential
+import com.google.firebase.auth.AuthResult
+import com.google.firebase.auth.FirebaseAuth
+import com.google.firebase.auth.FirebaseUser
+import com.google.firebase.auth.ProviderQueryResult
+import com.google.firebase.auth.SignInMethodQueryResult
+import com.nhaarman.mockito_kotlin.argumentCaptor
+import com.nhaarman.mockito_kotlin.mock
+import com.nhaarman.mockito_kotlin.times
+import com.nhaarman.mockito_kotlin.verify
+import com.nhaarman.mockito_kotlin.whenever
 import io.reactivex.observers.TestObserver
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.Before
